--- Machine.Specifications.ReSharper.Tests/Explorers/MspecTestMetadataExplorerTests.cs
@@ -1,10 +1,0 @@
-﻿using NUnit.Framework;
-
-namespace Machine.Specifications.ReSharper.Tests.Explorers
-{
-    [TestFixture]
-    public class MspecTestMetadataExplorerTests
-    {
-        
-    }
-}
